requirements.txt
Package             Version
------------------- ---------
asgiref             3.5.0
backports.zoneinfo  0.2.1
certifi             2021.10.8
charset-normalizer  2.0.12
click               8.0.3
Django              4.0.2
djangorestframework 3.13.1
filelock            3.5.1
huggingface-hub     0.4.0
idna                3.3
joblib              1.1.0
numpy               1.22.2
packaging           21.3
pip                 21.2.4
pyparsing           3.0.7
pytz                2021.3
PyYAML              6.0
regex               2022.1.18
requests            2.27.1
sacremoses          0.0.47
setuptools          58.0.4
six                 1.16.0
sqlparse            0.4.2
tokenizers          0.11.5
torch               1.10.2
tqdm                4.62.3
transformers        4.16.2
typing_extensions   4.1.1
urllib3             1.26.8
wheel               0.37.1
